requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
ansible==6.7.0
ansible-core==2.13.13
ansible-vault==2.1.0
anyio==4.6.0
attrs==24.2.0
base58==2.1.1
bittensor==8.0.0
bittensor-wallet==1.0.0
bt-decode==0.2.0a0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cytoolz==0.12.3
decorator==5.1.1
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==5.0.0
eth-utils==2.2.2
fastapi==0.110.3
frozenlist==1.4.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
Levenshtein==0.26.0
load-dotenv==0.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
msgpack-numpy-opentensor==0.5.0
multidict==6.1.0
munch==2.5.0
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.0.2
packaging==24.1
password-strength==0.0.3.post2
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
python-statemachine==2.3.6
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
resolvelib==0.8.1
retry==0.9.2
rich==13.8.1
scalecodec==1.2.11
sentry-sdk==2.15.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
substrate-interface==1.7.10
termcolor==2.4.0
toml==0.10.0
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
websocket-client==1.8.0
wheel==0.44.0
xxhash==3.5.0
yarl==1.13.1
